Update full PostController (#3)
* add vehicle command

* update vehicle request

* update vehicle controller

* update response code

* update repository

* update post controller

* fix include function of avtice post

* update controller

* update disableTracking

* update base bool response

+++ ClientService/ClientService.Application/User/Model/UserProfileResponse.cs
@@ -13,7 +13,7 @@ namespace ClientService.Application.User.Model
         public string Name { get; set; }
         public string Phone { get; set; }
         public string Avatar { get; set; }
-        public int AveragePoint { get; set; }
+        public float AveragePoint { get; set; }
         public bool IsUpdated { get; set; }
     }
 }
